// WADNR.Web/src/app/shared/generated/model/classification-upsert-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ClassificationUpsertRequest { 
    ClassificationSystemID?: number;
    DisplayName?: string | null;
    ClassificationDescription?: string | null;
    ThemeColor?: string | null;
    GoalStatement?: string | null;
    KeyImageFileResourceID?: number | null;
    ClassificationSortOrder?: number | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ClassificationUpsertRequestForm { 
    ClassificationSystemID?: FormControl<number>;
    DisplayName?: FormControl<string>;
    ClassificationDescription?: FormControl<string>;
    ThemeColor?: FormControl<string>;
    GoalStatement?: FormControl<string>;
    KeyImageFileResourceID?: FormControl<number>;
    ClassificationSortOrder?: FormControl<number>;
}

export class ClassificationUpsertRequestFormControls { 
    public static ClassificationSystemID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static DisplayName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ClassificationDescription = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ThemeColor = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static GoalStatement = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static KeyImageFileResourceID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ClassificationSortOrder = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}